Refactored homepage and moved utils to helper folder
